Backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
chardet==4.0.0
Django==5.1
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==2.10
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
requests==2.25.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
